# Use type-safe atomics where possible