<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

We already use [this library](https://github.com/uber-go/atomic) in [some tests](https://github.com/knative/serving/blob/44d8d27ce736d908bedbec7bd5674839dbba4d52/pkg/reconciler/autoscaling/kpa/kpa_test.go#L1594), so it's in go.mod already, and it's generally just as fast and more type-safe/harder to mess up accidentally.

Just to double-check, before/after of BenchmarkBreakerReserve:

~~~~
benchmark                                 old ns/op     new ns/op     delta
BenchmarkBreakerReserve/sequential-16     94.6          90.7          -4.12%
BenchmarkBreakerReserve/parallel-16       174           162           -6.90%

benchmark                                 old allocs     new allocs     delta
BenchmarkBreakerReserve/sequential-16     0              0              +0.00%
BenchmarkBreakerReserve/parallel-16       0              0              +0.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkBreakerReserve/sequential-16     0             0             +0.00%
BenchmarkBreakerReserve/parallel-16       0             0             +0.00%
~~~~

/assign @markusthoemmes @vagababov